requirements.txt
torch
torchvision
tqdm
matplotlib
numpy
pandas
scikit-learn
pillow
foolbox
eagerpy
